# swagger.yaml
openapi: 3.0.0
info:
  version: '1'
  title: FSARMY - MY BLOG API
  description: "This is the first professional api project we creating in full-stack army. This is a simple blog api application where anyone can create a new account, login to their system and create articles that will be visible publicly."
  contact:
    name: FS Army
    url: https://fsarmy.com
    email: support@fsarmy.com
servers: 
  - description: Dev
    url: http://localhost:4000/api/v1
  - description: Production
    url: https://app.fsamry.com/v1
...
paths:  
...
  "/articles": 
    get: 
      description: Retrive all published Articles
      tags: [Article]
      parameters: 
        - $ref: "#/components/parameters/pageParam"
        - $ref: "#/components/parameters/limitParam"
        - $ref: "#/components/parameters/sortTypeParam"
        - $ref: "#/components/parameters/sortByParam"
        - $ref: "#/components/parameters/searchParam"
      responses: 
        "200": 
          description: Return a list of article
          content: 
            "application/json": 
              schema: 
                type: object
                properties: 
                  data: 
                    type: array
                    items: 
                      allOf: 
                        - type: object
                          properties: 
                            id:
                              type: integer
                              format: int32
                              example: 1
                            title:
                              type: string
                              example: first article
                            cover:
                              type: string
                              format: url
                              example: https://s3.aws.com/myfirstblog/124.jpg
                            author:
                              $ref: "#/components/schemas/AuthorDTO"
                            link:
                              type: string
                              format: url
                              example: "/articles/1"
                          required: [id, title, author]
                        - $ref: "#/components/schemas/Timestamp"
                  pagination: 
                    $ref: "#/components/schemas/Pagination"
                  links: 
                    type: object
                    properties: 
                      self: 
                        type: string
                        format: url
                        example: "/articles?page=2&limit=10"
                      next:
                        type: string
                        format: url
                        example: "/articles?page=3&limit=10"
                      prev: 
                        type: string
                        format: url
                        example: "/articles?page=1&limit=10"
        "400": 
          $ref: "#/components/responses/400"
        "500":
          $ref: "#/components/responses/500"
    post: 
      operationId: createPost
      security: 
        - bearerAuth: []
      description: Create a new article
      tags: [Article]
      requestBody: 
        description: To create a new article user must need to provide valid data
        content: 
          "application/json": 
            schema: 
              type: object
              properties: 
                title:
                  type: string
                  example: First article
                body:
                  type: string
                  format: text
                  example: Body of the first article
                cover:
                  type: string
                  format: url
                  example: https://s3.aws.example.com/23nds
                status:
                  type: string
                  enum: [draft, published]
                  example: draft
              required: [title, body]  
      responses: 
        "201": 
          description: successfully created a new article
          content: 
            "application/json": 
              schema: 
                $ref: "#/components/schemas/ArticleCreateDTO"
        "400":
          $ref: "#/components/responses/400"
        "401": 
          $ref: "#/components/responses/401"
        "500":
          $ref: "#/components/responses/500"
...
components: 
  securitySchemes: 
    bearerAuth: 
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas: 
    Article: 
      allOf: 
        - type: object
          properties: 
            id: 
              type: integer
              format: int32
              example: 1
            authorId:
              type: integer
              format: int32
              example: 1
            title:
              type: string
              example: First Article
            body:
              type: string
              format: text
              example: The body of the first article
            cover:
              type: string
              format: url
              example: https://s3.aws.example.com/23nds
            status:
              type: string
              enum: [draft, published]
              example: draft
          required: [id, authorId, title, body, status]  
    ArticleCreateDTO:
      type: object
      properties:
        code:
          type: integer
          format: int32
          example: 201
        message:
          type: string
          example: "Article created successfully"
        data:
          $ref: "#/components/schemas/Article"
        links:
          type: object
          properties:
            self:
              type: string
              format: url
              example: "/articles/1"
            author:
              type: string
              format: url
              example: "/articles/1/author"
            comments:
              type: string
              format: url
              example: "/articles/1/comments"
...
    Timestamp:
      type: object
      properties:
        createdAt:
          type: string
          format: datetime
        updatedAt:
          type: string
          format: datetime
      required: [createdAt, updatedAt]
    AuthorDTO:
      type: object
      properties:
        id:
          type: integer
          format: int32
          example: 1
        name:
          type: string
          example: "A.B Raihan"
      required: [id, name]
    Pagination:
      type: object
      properties:
        page:
          type: integer
          format: int32
          example: 2
        limit:
          type: integer
          format: int32
          example: 10
        next:
          type: integer
          format: int32
          example: 3
        prev:
          type: integer
          format: int32
          example: 1
        totalPage:
          type: integer
          format: int32
          example: 5
        totalItems:
          type: integer
          format: int32
          example: 50
      required:
        - page
        - limit
        - totalPage
        - totalItems
  responses: 
    "500": 
      description: "Internal Server Error"
      content: 
        "application/json": 
          schema: 
            type: object
            properties: 
              code: 
                type: integer
                format: int32
                example: 500
              error: 
                type: string
                example: Internal Server Error
              message: 
                type: string
                example: We are sorry for the inconvenience. Please try again later.
    "400": 
      description: "Bad Request"
      content: 
        "application/json": 
          schema: 
            type: object
            properties: 
              code: 
                type: integer
                format: int32
                example: 400
              error: 
                type: string
                example: Bad Request
              data: 
                type: array
                items: 
                  type: object
                  properties: 
                    field:
                      type: string
                      example: email
                    message: 
                      type: string
                      example: Invalid email
                    in: 
                      type: string
                      enum: [query, body]
                      example: body
    "401": 
      description: Unauthorized
      content: 
        "application/json": 
          schema: 
            type: object
            properties: 
              code :
                type: integer
                format: int32
                example: 401
              error:
                type: string
                example: Unauthorized
              message: 
                type: string
                example: You don't have the right permissions
...
  parameters: 
    pageParam: 
      in: query
      name: page
      description: current page number
      schema:
        type: integer
        format: int32
        example: 2
    limitParam: 
      in: query
      name: limit
      description: maximum items to be returned
      schema: 
        type: integer
        format: int32
        example: 10
    sortTypeParam:
      in: query
      name: sort_type
      description: type of sort (asc or dsc)
      schema:
        type: string
        enum: [asc, dsc]
        example: dsc
    sortByParam:
      in: query
      name: sort_by
      description: Sort key
      schema:
        type: string
        example: updatedAt
    searchParam:
      in: query
      name: search
      description: Search term
      schema:
        type: string
        example: title
...